requirements.txt
tqdm==4.47.0
numpy==1.19.0
pandas==1.3.3
matplotlib==3.0.3
streamlit==1.3.0
plotly==5.4.0
pyproj==2.4.1
Pillow==8.4.0
